requirements.txt
Django==5.1.2
djangorestframework==3.15.2
celery==5.4.0
redis==5.2.0
django-crispy-forms==2.3
crispy-bootstrap5==2024.10
Pillow==10.4.0
django-import-export==4.3.3
django-celery-beat==2.7.0
dj-database-url==2.3.0
gunicorn==23.0.0
